feat: Remove region lock during session searches
The game supplies a list of 6 attributes when creating a matchmake
session or when searching for one; of which, the 5th attribute it
sends is a unique number identifying the game's region (0 for JPN,
1 for USA, 2 for EUR, etc).  This meant that players could only ever
match up with other players using the same regional copy of the game
(despite all regions connecting to the same SecureServer).

Ignoring this attribute during session search allows sessions
to be found regardless of anyone's game region.