feat: Ignore region lock during session searches
Zelda: Tri Force Heroes originally had a region lock put in place
that restricted matchmaking via Nintendo Network.

The game client supplies a list of 6 attributes when creating a matchmake
session or when searching for one.  Of which, the 5th attribute it sends
is a unique number identifying the game's region (0 for JPN, 1 for USA,
2 for EUR, etc).  Since all attribute values needed to match, it meant
that  players  could  only  ever  play  with  others  using  the  same
regional  copy  of  the  game  (despite  all  regional  copies
connecting to the same game server).

Ignoring this attribute during session search allows sessions to be
found regardless of anyone's game region.

I want to clarify that since the region lock was an intentional measure
taken during the game's development, its removal could lead to occasional
multiplayer desync issues in a few levels that feature automated dialogue,
as some of it may have inconsistent timings during cross-region play.

(I believe the desync issues can be addressed in the future through the
use of game-specific patches once those are developed.)